+++ src/FreddyConfig.cs
@@ -15,6 +15,22 @@ public class FreddyConfig : SyncedConfig<FreddyConfig>
     [DataMember] public SyncedEntry<bool> SOLO_GAMEPLAY { get; private set; } 
     [DataMember] public SyncedEntry<bool> RANDOM_SLEEP { get; private set; } 
     [DataMember] public SyncedEntry<int> TIME_BEFORE_LEAVING_SLEEP { get; private set; } 
+    
+    
+    
+    [DataMember] public SyncedEntry<int> BASE_SPAWN_CHANCES { get; private set; } 
+    [DataMember] public SyncedEntry<bool> USE_MOON_CHANCES { get; private set; } 
+    
+    [DataMember] public SyncedEntry<int> EXPERIMENTATION_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> ASSURANCE_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> VOW_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> OFFENSE_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> MARCH_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> REND_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> DINE_SPAWNRATE { get; private set; } 
+    [DataMember] public SyncedEntry<int> TITAN_SPAWNRATE { get; private set; } 
+    
+    
     // TODO : Create a setting for the distance between player for a player to be alone or not
     
     public FreddyConfig(ConfigFile cfg) : base("FreddyKrueger")
@@ -41,8 +57,40 @@ public class FreddyConfig : SyncedConfig<FreddyConfig>
             "ACTIVATE SOLO RUN MUST BE ON : This setting decides how much time will it take for you to automatically leave the dream world after entering it"
         );
         
+        BASE_SPAWN_CHANCES = cfg.BindSyncedEntry("Spawn Behaviour", "Default Spawn Chance in Percentage", 100,
+            "Option must be between 0 and 100 : This config changes the default spawn chances that Freddy manifest when you lend on a moon. This value is override on basic moons. This will be the value used on modded moons"
+        );
+        
+        USE_MOON_CHANCES = cfg.BindSyncedEntry("Spawn Behaviour", "Activate the per moon spawn rate", false,
+            "If set to true, the spawn rate will be different based on the moon difficulty. This will not work for non basic moons and will instead use the Default Spawn Chances in percentage"
+        );
+        
+        EXPERIMENTATION_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Experimentation", 14,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        ASSURANCE_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Assurance", 15,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        VOW_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Assurance", 20,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        OFFENSE_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Offense", 10,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        MARCH_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for March", 20,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        REND_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Rend", 30,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        
+        DINE_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Dine", 40,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
+        TITAN_SPAWNRATE = cfg.BindSyncedEntry("Moon Spawn rate", "Spawn rate for Titan", 25,
+            "This option is only activated if it Activate the per moon spawn rate is true. This value is in percentage %"
+        );
         
         
     }
-    
 }